chore: criando hook de produtores
