+++ src/hooks/useProdutores.js
@@ -1,0 +1,15 @@
+import { useEffect, useState } from "react";
+import { carregaProdutores } from "../services/carregaDados";
+
+export default function useProdutores() {
+    const [titulo, setTitulo] = useState('');
+    const [lista, setLista] = useState([]);
+
+    useEffect(() => {
+        const retorno = carregaProdutores();
+        setTitulo(retorno.titulo);
+        setLista(retorno.lista);
+    }, []);
+
+    return [titulo, lista];
+}
