+++ src/views/Home/components/Produtores.js
@@ -1,17 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { FlatList, Text, StyleSheet } from "react-native";
-import { carregaProdutores } from "../../../services/carregaDados";
+
 import Produtor from "./Produtor";
+import useProdutores from "../../../hooks/useProdutores";
 
 export default function Produtores({ topo: Topo }) {
-    const [titulo, setTitulo] = useState('');
-    const [lista, setLista] = useState([]);
-
-    useEffect(() => {
-        const retorno = carregaProdutores();
-        setTitulo(retorno.titulo);
-        setLista(retorno.lista);
-    }, []);
+    const [titulo, lista] = useProdutores();
 
     function TopoLista() {
         return <>
